Implementação de teste e correções

+++ src/screens/login/Login.js
@@ -28,21 +28,42 @@ class Login extends React.Component {
             window.location.reload();
         }
     }
+    verify = () =>{
+        const erro = [];
+    
+        if(!this.state.username){
+          erro.push("Campo matricula é obrigatorio");
+        }
+        if(!this.state.password  || this.state.password.trim() === ''){
+            erro.push("Campo senha é obrigatorio");
+          }
+        return erro;
+      }
 
     login = async() => {
+        const erro = this.verify();
+        if(erro.length > 0){
+          erro.forEach((message) =>{
+            showErrorMessage(message);
+          });
+          return false;
+        } 
         this.setState({ loading: true });
      
-       await this.context.login(this.state.username, this.state.password)
+        await this.context.login(this.state.username, this.state.password)
        .then(user =>
         {   
+          
             const msgs =  localStorage.getItem("token");
             const msg = JSON.parse(msgs);
+        
 
             if (!msgs.includes("detail")) {
                 showSuccessMessage(`${user.name}, você está logado!`);
                 this.props.history.push('/viewCommentsHome');
 
             } else {
+
                 localStorage.clear();
                 if(msgs === null){
                     return user.data;
@@ -54,12 +75,18 @@ class Login extends React.Component {
                 showErrorMessage(detalhe);                    
                 
             }
+           
 
         })
         .catch(error =>
         {
-            showErrorMessage('Servidor Indisponivel', error);
-            console.log(error);
+            if(localStorage.getItem("error") === "Request failed with status code 400"){
+                showErrorMessage("Número de Matrícula incorreto");     
+            } else{
+                showErrorMessage('Servidor Indisponivel', error);
+                console.log(error);
+            }
+            
         });
                             
         this.setState({ loading: false });
